requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
idna==2.8
oauth2client==4.1.3
oauthlib==3.1.0
pipenv==2018.11.26
pyasn1==0.4.6
pyasn1-modules==0.2.6
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
soupsieve==1.9.3
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.7.3
virtualenv-clone==0.5.3
wincertstore==0.2
